% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{create_keyring}
\alias{create_keyring}
\title{Create a new keyring}
\usage{
create_keyring(ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{ring}{Optional. Name of keyring.}
}
\value{
NULL (invisibly)
}
\description{
Make a place for R to remember the user's passwords.
}
\details{
Function includes a check for \code{\link{has_keyring}}.
}
\examples{
\dontrun{
create_keyring()
}
}
