% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{currently_testing}
\alias{currently_testing}
\title{Check if currently running automated tests}
\usage{
currently_testing()
}
\value{
Logical
}
\description{
Unexported. Checks if currently running automated tests
with the \code{devtools} / \code{testthat} suite.
}
\details{
If I don't check for this, technically I"m still in an interactive session
even if I'm running automated tests from within RStudio.
}
