% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{drop_key}
\alias{drop_key}
\title{Forget a key}
\usage{
drop_key(key, ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{key}{Character. The system for which yser wants to retrieve a password.}

\item{ring}{Optional. Name of keyring.}
}
\value{
NULL (invisibly)
}
\description{
Delete the password for a given system.
}
\details{
Function includes checks for \code{\link{has_keyring}},
\code{\link{keyring_locked}}, and \code{\link{has_key}}.
}
\examples{
\dontrun{
drop_key("snowflake")
}
}
