% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{drop_keyring}
\alias{drop_keyring}
\title{Drop the keyring}
\usage{
drop_keyring(ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{ring}{Optional. Name of keyring.}
}
\value{
NULL (invisibly)
}
\description{
Delete the keyring, the place the user has set up to remember passwords.
}
\details{
Usually useful only if the user has forgotten their master password.

Function includes a check for \code{\link{has_keyring}}.
}
\examples{
\dontrun{
drop_keyring()
}
}
