% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{has_keyring}
\alias{has_keyring}
\title{Check for keyring}
\usage{
has_keyring(ring = getOption("speak.keyring", "friend"))
}
\arguments{
\item{ring}{Optional. Name of keyring.}
}
\value{
Logical
}
\description{
Is the user set up for R to remember their passwords?
}
\examples{
\dontrun{
has_keyring()
}
}
