% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interactive_session}
\alias{interactive_session}
\title{Check for interactive session}
\usage{
interactive_session()
}
\value{
Logical
}
\description{
Unexported. Is the function being called in an interactive setting?
}
\details{
\code{\link[base]{interactive}} has thrown errors for me in the past.
This way all such errors are abstracted out to a single function.

It also allows me to check for \code{\link{currently_testing}}.
}
