% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_locked}
\alias{keyring_locked}
\title{Check if keyring is locked}
\usage{
keyring_locked(ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{ring}{Optional. Name of keyring.}
}
\value{
Logical
}
\description{
Does the user need to enter a master password?
}
\details{
Function includes a check for \code{\link{has_keyring}}.
}
\examples{
\dontrun{
keyring_locked()
}
}
