% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{list_keys}
\alias{list_keys}
\title{List keys}
\usage{
list_keys(key = NULL, ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{key}{Character (optional): key name. Chiefly for use in error messages,
as this function is also called by \code{\link{has_key}},
which in turn is called by \code{\link{set_key}}, \code{\link{get_key}},
and \code{\link{drop_key}}.}

\item{ring}{Optional. Name of keyring.}
}
\value{
Character vector of key names.
}
\description{
What all passwords has the user set?
}
\details{
Function includes checks for \code{\link{has_keyring}}
and \code{\link{keyring_locked}}.
}
\examples{
\dontrun{
list_keys()
}
}
