% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{lock_keyring}
\alias{lock_keyring}
\title{Lock the keyring}
\usage{
lock_keyring(ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{ring}{Optional. Name of keyring.}
}
\value{
NULL (invisibly)
}
\description{
Hide passwords from other users until master password is entered later.
}
\details{
Function includes checks for \code{\link{has_keyring}}
and \code{\link{keyring_locked}}.
If the keyring is already locked,
\code{\link{lock_keyring}} will message the user.
}
\examples{
\dontrun{
lock_keyring()
}
}
