% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{offer_keyring}
\alias{offer_keyring}
\title{Offer to set up keyring}
\usage{
offer_keyring(ring = getOption("speak.keyring", "friend"))
}
\arguments{
\item{ring}{Optional. Name of keyring.}
}
\value{
Logical. Does the user want to set up a master password?
}
\description{
Unexported, used in \code{\link{keyring_is_locked}}
and \code{\link{unlock_keyring}}.
}
\details{
Will set up keyring if user accepts, or fail informatively otherwise.
}
