% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{unlock_keyring}
\alias{unlock_keyring}
\title{Unlock the keyring}
\usage{
unlock_keyring(master_password = NULL, key = NULL,
  ring = getOption("keyring_keyring", "friend"))
}
\arguments{
\item{master_password}{Character, to specify master password in plain text,
or NULL to type it in at runtime.}

\item{key}{Character (optional): key name. Chiefly for use in error messages,
as this function is also called by \code{\link{list_keys}},
which is called either directly or indirectly by \code{\link{has_key}},
\code{\link{set_key}}, \code{\link{get_key}}, and \code{\link{drop_key}}.}

\item{ring}{Optional. Name of keyring.}
}
\value{
NULL (invisibly)
}
\description{
Enter the master password for R to remember all other passwords.
}
\details{
\strong{It is recommended to enter your master password interactively}
using \code{unlock_keyring()}.
You can also use \code{unlock_keyring("your_master_password")}
to enter your master password programmatically,
but \strong{entering passwords in plain text is not recommended}.
You could forget and save it as a script,
and either way it will show up in your \code{.Rhistory} file.

Function includes checks for \code{\link{has_keyring}}
and \code{\link{keyring_locked}}.
If the keyring is already unlocked,
\code{\link{unlock_keyring}} will message the user.
}
\examples{
\dontrun{
unlock_keyring()
}
}
